re-organized score models
