+++ src/treeffuser/_score_models/__init__.py
@@ -1,0 +1,7 @@
+from treeffuser._score_models._lgbm_score_model import LightGBMScore
+from treeffuser._score_models._base import Score
+
+__all__ = [
+    "Score",
+    "LightGBMScore"
+]
